Fix v8 external snapshot code for Windows.
Needed to land https://codereview.chromium.org/826083004/.

BUG=421063, 439661
R=tsepez@chromium.org

Review URL: https://codereview.chromium.org/848443002

Patch from Ross McIlroy <rmcilroy@chromium.org>.

+++ samples/pdfium_test.cc
@@ -105,7 +105,7 @@ static bool GetExternalData(const Options& options,
     return false;
   }
   result_data->data = const_cast<const char*>(data_buffer);
-  result_data->raw_size = data_length;
+  result_data->raw_size = static_cast<int>(data_length);
   return true;
 }
 #endif  // V8_USE_EXTERNAL_STARTUP_DATA
